% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{calculate_bis}
\alias{calculate_bis}
\title{Calculate the bisilhouette score without repeats}
\usage{
calculate_bis(data, row_clustering, col_clustering, method = "euclidean")
}
\arguments{
\item{data}{data matrix, shape (N, p).}

\item{row_clustering}{binary matrix indicating row clustering,
shape(N, K).}

\item{col_clustering}{binary matrix indicating column clustering,
shape(p, K).}

\item{method}{distance metric to use, str. Default is "euclidean".}
}
\value{
list containing;
- bisil: bisilhouette score, float.
- vals: individual sample scores, shape (N, ).
- repeat: if TRUE, a random row cluster has been added
and repeats are needed, bool.
}
\description{
Calculate the bisilhouette score without repeats
}
