% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{calculate_scores}
\alias{calculate_scores}
\title{Calculate the a and b values for the kth bicluster.}
\usage{
calculate_scores(distances, indices, k, n_clusts_row, clust_two)
}
\arguments{
\item{distances}{distance (over the columns in question) matrix,
shape (N, N).}

\item{indices}{binary vector indicating row cluster, shape (N, ).}

\item{k}{current row cluster index, int.}

\item{n_clusts_row}{number of row clusters, int.}

\item{clust_two}{binary matrix indicating row clustering,
shape(N, n_clusts_row).}
}
\value{
bisil: bisilhouette score, float.
}
\description{
Calculate the a and b values for the kth bicluster.
}
