% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{check_unique}
\alias{check_unique}
\title{Check if there are at least two unique row clusters.}
\usage{
check_unique(row_clustering)
}
\arguments{
\item{row_clustering}{binary matrix indicating row clustering,
shape(N, k).}
}
\value{
TRUE if there are not at least three unique row clusters, bool.
}
\description{
Check if there are at least two unique row clusters.
}
