% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.r
\name{df_plot}
\alias{df_plot}
\title{Data frame plot}
\usage{
df_plot(values)
}
\arguments{
\item{values}{A list of bisilhouette scores.}
}
\value{
A data frame for plotting.
}
\description{
Function to create a data frame for
plotting from bisilhouette coefficients.
}
