% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.r
\name{x_breaks}
\alias{x_breaks}
\title{X breaks}
\usage{
x_breaks(df)
}
\arguments{
\item{df}{A data frame of bisilhouette scores.}
}
\value{
A vector of breaks for the x-axis.
}
\description{
Function to determine the breaks for the x-axis.
}
