% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.r
\name{y_breaks}
\alias{y_breaks}
\title{Break function}
\usage{
y_breaks(lower)
}
\arguments{
\item{lower}{A vector of the lower values of the bisilhouette scores.}
}
\value{
A vector of breaks for the y-axis.
}
\description{
Function to determine the breaks for the y-axis.
}
