% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{bisilhouette}
\alias{bisilhouette}
\title{Calculate the bisilhouette score.}
\usage{
bisilhouette(
  data,
  row_clustering,
  col_clustering,
  method = "euclidean",
  seed = TRUE,
  n_reps = 10
)
}
\arguments{
\item{data}{data matrix, shape (N, p).}

\item{row_clustering}{binary matrix indicating row clustering,
shape(N, k).}

\item{col_clustering}{binary matrix indicating column clustering,
shape(p, k).}

\item{method}{distance metric to use, str. Default is "euclidean".}

\item{seed}{seed if seed should be set for random number generation, int.
Default is FALSE.}

\item{n_reps}{number of repetitions if random biclusters added, int.
Default is 10.}
}
\value{
list containing;
- bisil: bisilhouette score, float.
- vals: individual sample scores, shape (N, ).
}
\description{
Calculate the bisilhouette score.
}
\examples{
data <- matrix(stats::rnorm(50), nrow = 10)
row_clustering <- cbind(
  stats::rbinom(10, 1, 0.5),
  stats::rbinom(10, 1, 0.5),
  stats::rbinom(10, 1, 0.5)
)
col_clustering <- cbind(
  stats::rbinom(5, 1, 0.5),
  stats::rbinom(5, 1, 0.5),
  stats::rbinom(5, 1, 0.5)
)
bisilhouette(data, row_clustering, col_clustering)
}
